Fix Jinja2 variables quoting in shell
Merely enclosing a value with "" is not sufficient if the value contains
problematic characters (most notably $, ` or ").

Instead of assuming things about the value, use Jinja2's built-in shell
quoting filter [1], which quotes and/or escapes strings as necessary for
use in shell scripts.

Also introduce the $repo_name variable, to reduce the amount of
variables templating required (and thus make the script slightly more
readable).

[1] https://docs.ansible.com/ansible/latest/user_guide/playbooks_filters.html#manipulating-strings